refactor(network): 简化代码并消除范式问题
- 移除 HttpContext 转接层，Connection 直接持有 HttpRequest
- Channel: handleEvent EPOLLOUT 优先，回调后立即返回防 use-after-free
- Acceptor: 初始化器直接传入 listenSocket_.fd()，消除 Channel 重赋值
- 删除 EventLoopThread::exiting_、Socket::close() 等死代码
- HttpResponse::sendResponse 参数改为 const 引用
- HttpRequest::contentLength_ 改为 size_t，消除类型转换
- Socket.h 精简头文件，实现所需的头文件移至 .cpp
- main_stage2.cpp: onRead 增加 n==0 和 EAGAIN 处理

+++ network/src/HttpRequest.cpp
@@ -63,7 +63,7 @@ bool HttpRequest::parse(const char* data, size_t len) {
             break;
         }
         case kExpectBody: {
-            if (buf_.size() >= static_cast<size_t>(contentLength_)) {
+            if (buf_.size() >= contentLength_) {
                 body_ = buf_.substr(0, contentLength_);
                 buf_.erase(0, contentLength_);
                 state_ = kGotAll;
